Use CDN for bootstrap voodoo
Since we're not actually diverging from the upstream bootstrap voodoo
(confirmed with diff -Nur), we may as well load it from a CDN.  But
add checksums to the <link> tags so the CDN won't serve us bullshit.

While here: tidy up _includes/head.html and _includes/footer.html to
more closely match the suggested order of things in the bootstrap and
Jekyll documentation.