Remove dependency on google fonts.
TL;DR serve fonts from hillhacks.in server.

The fonts are imported from google fonts archive[1], and this change
removes client call to google servers.

The fonts themselves have permissive licenses afiu, of which you can
find in the respective folders.

In future, if one wants to import a google font, get it from [1] and
add the @font-face bits to _sass/_fonts.scss

[1] https://github.com/google/fonts